requirements.txt
pygame-ce==2.5.3
pygame_gui==0.6.14
python-i18n==0.3.9
